+++ packages/build-tools/plugins/sass-export-data/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.7.0](https://github.com/basaltinc/theme-tools/tree/master/packages/sass-export-data/compare/v2.6.0...v2.7.0) (2019-09-13)
+
+**Note:** Version bump only for package @bolt/sass-export-data
+
+
+
+
+
 ## [2.5.1](https://github.com/basaltinc/theme-tools/tree/master/packages/sass-export-data/compare/v2.5.0...v2.5.1) (2019-06-21)
 
 **Note:** Version bump only for package @bolt/sass-export-data