+++ packages/components/bolt-stack/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.7.0](http://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-stack/compare/v2.6.0...v2.7.0) (2019-09-13)
+
+**Note:** Version bump only for package @bolt/components-stack
+
+
+
+
+
 # [2.6.0](http://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-stack/compare/v2.6.0-beta.2...v2.6.0) (2019-08-30)
 
 **Note:** Version bump only for package @bolt/components-stack